@@ -1,15 +1,19 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-var ts = require("typescript");
-function isJsxOpeningLike(node) {
-    return (node.kind === ts.SyntaxKind.JsxOpeningElement ||
-        node.kind === ts.SyntaxKind.JsxSelfClosingElement);
-}
+var ts = require("byots");
 function isDefineMessages(el, tagName) {
-    return (el.kind === ts.SyntaxKind.VariableDeclaration &&
+    return (ts.isVariableDeclaration(el) &&
         el.initializer &&
+        ts.isCallExpression(el.initializer) &&
         el.initializer.expression &&
+        ts.isIdentifier(el.initializer.expression) &&
         el.initializer.expression.text === tagName);
+}
+function emptyObject(obj) {
+    for (var x in obj) {
+        return false;
+    }
+    return true;
 }
 function findProps(node) {
     var res = [];
@@ -18,13 +22,19 @@ function findProps(node) {
         if (!node) {
             return undefined;
         }
-        if (node.kind === ts.SyntaxKind.ObjectLiteralExpression) {
+        if (ts.isObjectLiteralExpression(node)) {
             node.properties.forEach(function (p) {
-                var props = {};
                 var prop = {};
-                if (p.initializer.properties) {
+                if (ts.isPropertyAssignment(p) &&
+                    ts.isObjectLiteralExpression(p.initializer) &&
+                    p.initializer.properties) {
                     p.initializer.properties.forEach(function (ip) {
-                        prop[ip.name.text] = ip.initializer.text;
+                        if (ts.isIdentifier(ip.name)) {
+                            var name_1 = ip.name.text;
+                            if (ts.isPropertyAssignment(ip) && ts.isStringLiteral(ip.initializer)) {
+                                prop[name_1] = ip.initializer.text;
+                            }
+                        }
                     });
                     res.push(prop);
                 }
@@ -41,12 +51,12 @@ function findFirstJsxOpeningLikeElementWithName(node, tagName, dm) {
         if (!node) {
             return undefined;
         }
-        if (dm && node.getNamedDeclarations) {
+        if (dm && ts.isSourceFile(node)) {
             var nd = node.getNamedDeclarations();
-            nd.forEach(function (element) {
+            nd.forEach(function (element, key) {
                 element.forEach(function (el) {
                     if (isDefineMessages(el, tagName)) {
-                        if (el.initializer.kind === ts.SyntaxKind.CallExpression &&
+                        if (ts.isCallExpression(el.initializer) &&
                             el.initializer.arguments.length) {
                             var nodeProps = el.initializer.arguments[0];
                             var props = findProps(nodeProps);
@@ -57,8 +67,8 @@ function findFirstJsxOpeningLikeElementWithName(node, tagName, dm) {
             });
         }
         else {
-            if (isJsxOpeningLike(node) &&
-                node.tagName.kind === ts.SyntaxKind.Identifier) {
+            if (ts.isJsxOpeningLikeElement(node) &&
+                ts.isIdentifier(node.tagName)) {
                 var childTagName = node.tagName;
                 if (childTagName.text === tagName) {
                     res.push(node);
@@ -73,7 +83,6 @@ function main(contents) {
     var sourceFile = ts.createSourceFile("file.ts", contents, ts.ScriptTarget.ES2015, false, ts.ScriptKind.TSX);
     var elements = findFirstJsxOpeningLikeElementWithName(sourceFile, "FormattedMessage");
     var dm = findFirstJsxOpeningLikeElementWithName(sourceFile, "defineMessages", true);
-    var emptyObject = function (o) { return JSON.stringify(o) === "{}"; };
     var res = elements
         .map(function (element) {
         var msg = {};
